allow OIDC for WP accounts to be optional
Previously, using OIDC to log in to WordPress accounts could be
on or off.  When it was off, the WordPress login form would
display a "Log in with Single Sign On" button that would log
the user in to the website but not log them into WordPress,
and also the normal WordPress login form that would log them in
to WordPress but not into the website/OIDC.  This was confusing
and hard to explain; users expected the two login methods
on the WordPress login page to do the same thing.

This patch changes using OIDC to log in to WordPress accounts
to have three settings: no, optional, and yes:

* "no" displays only the standard WordPress login form without
the SSO button.

* "optional" allows users to choose between OIDC and their
WordPress credentials each time they log in.  If a user chooses
OIDC but does not have a WordPress user account, they will
receive an "Access Denied" error.

* "yes" automatically logs the user in to both the website
via OIDC as well as any corresponding WordPress account.  If the
user does not have a WordPress account, they will receive an
"Access Denied" error.

These changes affect only logging in to WordPress via
/wp-login.php, /wp-admin, or other methods that use a WordPress
login URL.  The changes do not affect users logging in to the
website using a UMich OIDC Login plugin shortcode / URL.